# swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/profile": {
      "post": {
        "operationId": "ProfileController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProfileDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": ["Profile"]
      }
    },
    "/api/profile/{id}": {
      "patch": {
        "operationId": "ProfileController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProfileDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": ["Profile"]
      }
    },
    "/api/profile/{userId}": {
      "delete": {
        "operationId": "ProfileController_deleteByUserId",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": ["Profile"]
      }
    },
...
    "/api/transportings": {
      "post": {
        "operationId": "TransportingsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTransportingDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Transporting record created successfully"
          },
          "400": {
            "description": "Invalid input"
          }
        },
        "summary": "Create a new transporting record",
        "tags": ["transportings"]
      },
      "patch": {
        "operationId": "TransportingsController_update",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTransportingDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Transporting record updated successfully"
          },
          "404": {
            "description": "Transporting record not found"
          }
        },
        "summary": "Update a transporting record",
        "tags": ["transportings"]
      }
    },
    "/api/transportings/{id}": {
      "get": {
        "operationId": "TransportingsController_findById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Transporting ID",
            "schema": {
              "example": "507f1f77bcf86cd799439011",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Transporting record found"
          },
          "404": {
            "description": "Transporting record not found"
          }
        },
        "summary": "Get transporting record by ID",
        "tags": ["transportings"]
      }
    },
    "/api/transportings/order/{orderId}": {
      "get": {
        "operationId": "TransportingsController_findByOrderId",
        "parameters": [
          {
            "name": "orderId",
            "required": true,
            "in": "path",
            "description": "Order ID",
            "schema": {
              "example": "507f1f77bcf86cd799439012",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Transporting record found"
          },
          "404": {
            "description": "Transporting record not found"
          }
        },
        "summary": "Get transporting record by order ID",
        "tags": ["transportings"]
      }
    },
    "/api/transportings/company/{companyId}": {
      "get": {
        "operationId": "TransportingsController_findByCompanyId",
        "parameters": [
          {
            "name": "companyId",
            "required": true,
            "in": "path",
            "description": "Company ID",
            "schema": {
              "example": "507f1f77bcf86cd799439013",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of transporting records found"
          },
          "404": {
            "description": "No transporting records found for the company"
          }
        },
        "summary": "Get transporting records by company ID",
        "tags": ["transportings"]
      }
    },
    "/api/transportings/{id}/cancel": {
      "patch": {
        "operationId": "TransportingsController_cancel",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Transporting ID",
            "schema": {
              "example": "507f1f77bcf86cd799439011",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Transporting record canceled successfully"
          },
          "400": {
            "description": "Transporting record cannot be canceled"
          },
          "404": {
            "description": "Transporting record not found"
          }
        },
        "summary": "Cancel a transporting record",
        "tags": ["transportings"]
      }
    },
...
  },
  "info": {
    "title": "practice",
    "description": "API documentation for my practice app",
    "version": "0.0.1",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "Cart": {
        "type": "object",
        "properties": {}
      },
      "CreateProfileDto": {
        "type": "object",
        "properties": {
          "phoneNumber": {
            "type": "string",
            "description": "Iranian phone number of the user",
            "example": "+989123456789"
          },
          "nationalId": {
            "type": "string",
            "description": "National ID of the user",
            "example": "1234567890"
          },
          "firstName": {
            "type": "string",
            "description": "First name of the user",
            "example": "Ali"
          },
          "lastName": {
            "type": "string",
            "description": "Last name of the user",
            "example": "Hosseini"
          },
          "address": {
            "type": "string",
            "description": "Address of the user",
            "example": "Tehran, Valiasr St., No. 123"
          },
          "walletId": {
            "type": "string",
            "description": "Wallet ID associated with the user (optional)",
            "example": "507f1f77bcf86cd799439011"
          },
          "orders": {
            "description": "List of order IDs associated with the user",
            "example": ["507f1f77bcf86cd799439011", "507f1f77bcf86cd799439012"],
            "default": [],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "transactions": {
            "description": "List of transaction IDs associated with the user",
            "example": ["507f1f77bcf86cd799439013", "507f1f77bcf86cd799439014"],
            "default": [],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "favorites": {
            "description": "List of favorite product IDs",
            "example": ["507f1f77bcf86cd799439015", "507f1f77bcf86cd799439016"],
            "default": [],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "cart": {
            "description": "User’s shopping cart",
            "default": [],
            "allOf": [
              {
                "$ref": "#/components/schemas/Cart"
              }
            ]
          }
        },
        "required": [
          "phoneNumber",
          "nationalId",
          "firstName",
          "lastName",
          "address",
          "orders",
          "transactions",
          "favorites",
          "cart"
        ]
      },
      "UpdateProfileDto": {
        "type": "object",
        "properties": {
          "phoneNumber": {
            "type": "string",
            "description": "Iranian phone number of the user",
            "example": "+989123456789"
          },
          "nationalId": {
            "type": "string",
            "description": "National ID of the user",
            "example": "1234567890"
          },
          "firstName": {
            "type": "string",
            "description": "First name of the user",
            "example": "Ali"
          },
          "lastName": {
            "type": "string",
            "description": "Last name of the user",
            "example": "Hosseini"
          },
          "address": {
            "type": "string",
            "description": "Address of the user",
            "example": "Tehran, Valiasr St., No. 123"
          },
          "walletId": {
            "type": "string",
            "description": "Wallet ID associated with the user (optional)",
            "example": "507f1f77bcf86cd799439011"
          },
          "orders": {
            "description": "List of order IDs associated with the user",
            "example": ["507f1f77bcf86cd799439011", "507f1f77bcf86cd799439012"],
            "default": [],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "transactions": {
            "description": "List of transaction IDs associated with the user",
            "example": ["507f1f77bcf86cd799439013", "507f1f77bcf86cd799439014"],
            "default": [],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "favorites": {
            "description": "List of favorite product IDs",
            "example": ["507f1f77bcf86cd799439015", "507f1f77bcf86cd799439016"],
            "default": [],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "cart": {
            "description": "User’s shopping cart",
            "default": [],
            "allOf": [
              {
                "$ref": "#/components/schemas/Cart"
              }
            ]
          }
        }
      },
...
      "CreateTransportingDto": {
        "type": "object",
        "properties": {
          "orderId": {
            "type": "string",
            "description": "MongoDB ObjectId of the order associated with the shipment",
            "example": "507f1f77bcf86cd799439011"
          },
          "companyId": {
            "type": "string",
            "description": "MongoDB ObjectId of the company handling the shipment",
            "example": "507f1f77bcf86cd799439012"
          },
          "carrier": {
            "type": "string",
            "description": "Name of the carrier",
            "example": "DHL Express"
          },
          "trackingNumber": {
            "type": "string",
            "description": "Tracking number for the shipment",
            "example": "123456789"
          },
          "status": {
            "type": "string",
            "description": "Status of the shipment",
            "enum": ["pending", "shipped", "delivered", "canceled"],
            "example": "shipped"
          },
          "estimatedDelivery": {
            "format": "date-time",
            "type": "string",
            "description": "Estimated delivery date for the shipment",
            "example": "2025-06-10"
          }
        },
        "required": ["orderId", "companyId", "carrier", "trackingNumber"]
      },
      "UpdateTransportingDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "MongoDB ObjectId of the transporting record",
            "example": "507f1f77bcf86cd799439011"
          }
        },
        "required": ["id"]
      },
...
    }
  }
}